mlaugeas: Resynch with upstream ocaml-augeas.
Resynch with ocaml-augeas to 1cf5aef99b26a46529ca797547c0b49627fffe78
(version 0.6).

This includes numerous upstream enhancements to the API, but most
critically includes a fix which is required for OCaml 4.09.0 because
caml_named_value now returns a const pointer.

This adds the license file from ocaml-augeas, although it is not
strictly needed since this license is compatible with libguestfs, but
may be better to be explicit about it.