utils: split string list functions in own file
Few of them are used also when the whole libutils.a is not needed (or
usable), also because the whole utils.c indirectly pulls gnulib.

As solution, split them in an own file, with an own header so they can
be used alone.  To maintain compatibility with users of the whole
libutils.a, include the new header in the existing guestfs-utils.h.