utils: conditionally include config.h on HAVE_CONFIG_H
autotools define HAVE_CONFIG_H to signal a config.h was generated; since
this file is built directly also in bindings without the use of
config.h, make sure config.h is not requested in those cases.

This file uses only POSIX functions anyway, so it should not be a
problem.