ocaml: Use caml_alloc_initialized_string instead of memcpy.
Since about 2017 OCaml has had a function for creating an initialized
string.  This uses the function instead of caml_alloc_string + memcpy
(which doesn't work for OCaml 4.10) and defines a replacement if the
function is missing.

Note this requires configure.ac in libguestfs.git and virt-v2v.git to
define HAVE_CAML_ALLOC_INITIALIZED_STRING.