options: Compile blocksize code conditionally.
Since the blocksize option was added since the minimum version of
libguestfs for virt-v2v, we need to make it conditional so that
virt-v2v can still be compiled with the older libguestfs.

This commit does the minimum fix to make it build.  It doesn't
conditionalize the --blocksize option altogether.  However I believe,
at least for virt-v2v, that is OK since the --blocksize option is not
used there.

Thanks: Pino Toscano
Fixes: commit c33e0036c70ce68d40df92e4a6c0423e136e005c