common: Replace -I$(top_srcdir)/include with $(INCLUDE_DIRECTORY).
We can only use the top-level include directory when we're compiling
libguestfs itself, not other packages such as virt-v2v that also use
the common submodule.  So therefore we need to turn this into a macro
which can be defined only when compiling libguestfs and will be empty
for other packages.