mlpcre: Fix error value when PCRE.sub references unset capture.
The underlying function actually returns PCRE2_ERROR_UNSET in this
case.

PCRE2_ERROR_NOSUBSTRING is returned for a capture number which is
wrong.  This suggests we ought to raise PCRE.Error (instead of
Not_found) for this case.  However there is an existing unit test:

  let re2 = compile "(a+)(b*)" in
  ..
  (try ignore (sub 3) with Not_found -> ());

which suggests that the old code would raise Not_found in this case,
so I don't change the behaviour here.

Fixes: commit 0b244f2b0cfc920f2432f5df45e108583f0d2901