Makefile.am: supply missing $(LIBGUESTFS_CFLAGS)
C source code in the "edit", "mlvisit", "parallel", "progress", "structs",
"visit", and "windows" modules includes "guestfs.h", but the makefiles of
those modules do not add $(LIBGUESTFS_CFLAGS) to the respective module
CFLAGS macros.

This is a problem when these modules are built as a part of guestfs-tools,
against a just-built libguestfs tree, as follows:

> guestfs-tools$ ../libguestfs/run ./configure
> guestfs-tools$ ../libguestfs/run make

Example error:

>   CC       libmlvisit_a-visit-c.o
> visit-c.c:33:10: fatal error: guestfs.h: No such file or directory
>    33 | #include "guestfs.h"
>       |          ^~~~~~~~~~~

Add the missing $(LIBGUESTFS_CFLAGS) instances.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20210906114226.14689-2-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>