mltools: Add a new On_exit module to handle exit actions
OCaml has a Stdlib.at_exit function (similar to C atexit), but it
doesn't run if the program is interrupted by a signal.  It's also
rather awkward to use for simple actions like making sure a temporary
file is deleted on exit.

For a long time we've have a couple of functions -- unlink_on_exit and
rmdir_on_exit -- to handle the actions of cleaning up temporary files,
and we've used signals and at_exit in an ad-hoc way elsewhere.  This
module provides a more complete solution.