mltools: Make On_exit safe if exit actions throw exceptions
If one of the exit actions throws an exception then the rest of the
actions would not run (except for unlink which was already protected
by a try clause).  Modify the function that runs the exit actions so
it turns exceptions into debug output and continues running the rest
of the actions.