utils: Fix usage of strerror_r
  $ ./run guestfish -vx add-drive foo "readonly:true"
  libguestfs: trace: set_pgroup true
  libguestfs: trace: set_pgroup = 0
  libguestfs: trace: add_drive "foo" "readonly:true"
  libguestfs: error: foo:
  libguestfs: trace: add_drive = -1 (error)
  libguestfs: trace: close
  libguestfs: closing guestfs handle 0x55c0bacf12a0 (state 0)

Note the "error: foo: " line.  After hexdumping this I found that it
is actually printing random rubbish after the colon.  It turns out
that strerror_r was not updating the buffer and so we were printing
random bytes from the stack.

strerror_r is hard to use and worse still there are two incompatible
variants in glibc.  Close reading of the GNU variant that we are using
shows that it doesn't always use the buffer passed in, which explains
what was happening here.

Provide an easier to use wrapper around strerror_r and try to cope
with the two variants using _GNU_SOURCE.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2030396
Reported-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>
Tested-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>