Tools_utils.wrap: only wrap text for TTYs
Whenever stdout and/or stderr is redirected (usually to a regular file),
users mostly do that to capture logs, for later searching. Wrapping
interferes with grep though, so make "wrap" just print the string when the
output channel is not a TTY. The requested indentation is honored though.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1820221
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220211153225.17424-3-lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>