mltools: Reimplement On_exit to use a list of actions
Previously we used separate lists of files, dirs, pids, etc.  This
makes it harder to introduce new features to reorder actions.
Reimplement the module so we use a simple list of actions, where each
action can have type File, Rm_rf, Kill, etc.  Iterate through this
list on exit to execute the actions.

The actions will run in a different order from before, but we didn't
guarantee the ordering before.  Apart from that the functionality is
unchanged.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>