options: Don't attempt to scan LVs if "lvm2" feature is not available
Since we added support for transparent decryption of LUKS in
commit a232e62dcf ("fish: '-i' option automatically handles whole-disk
encryption") we have always assumed that the "lvm2" feature is
available and so we could call guestfs_vg_activate_all.  That API
would have failed in the unlikely event that the feature is not
available, preventing opening any (even unencrypted) disk.

The code has changed greatly since then but the basic problem remains.
Test for the "lvm2" feature, otherwise avoid trying to scan LVs.

Reported-by: Feng Li
Fixes: a232e62dcf
Reviewed-by: Laszlo Ersek <lersek@redhat.com>