options: Free key_store after use
The lifetime of the struct key_store is very unclear because it is
allocated the first time you call key_store_import_key.  But it seems
that the intention from existing code is that free_key_store ought to
free the key_store struct as well as the keys inside.

We ought to clean this up further by adding an explicit init function,
but this fixes a leak detected by valgrind.

Fixes: commit f84d95474ccf873a09eb0df9b54554a2d84a886b