utils: Add function for copying the environment and adding new entries
libguestfs is currently calling setenv at an unsafe location between
fork and exec.  To fix this we need a way to copy and modify the
environment before fork and then we can pass the modified environ to
execve-like functions.  nbdkit already does the same so use that code.

This function is copied and adapted from here under a compatible license:
https://gitlab.com/nbdkit/nbdkit/-/blob/master/common/utils/environ.c