mlcustomize: Allow virtio-win handle from path or libosinfo
For compatibility with virt-v2v we previously only allowed a
virtio-win handle to be created out of the $VIRTIO_WIN environment
variable.

Using environment variables for configuration doesn't work for
virt-customize (and in fact isn't a great idea for virt-v2v either).
Add two new methods for creating the handle, either from a path (to
ISO or directory), or "from libosinfo".  The latter means to use
libosinfo to get the drivers, and not use virtio-win explicitly.

In virt-customize these will map to:

  virt-customize --inject-virtio-win /path/to/virtio-win.iso
  virt-customize --inject-virtio-win /path/to/virtio-win/
  virt-customize --inject-virtio-win "osinfo"