mldrivers: Add support for reading Windows drivers from DriverDatabase
Add support for reading the Windows drivers from HKLM\SYSTEM\DriverDatabase
in the Windows Registry.  This will only work for Windows >= 7, not
for older versions which used the CriticalDeviceDatabase.

It returns a list of the drivers installed in the guest, and the USB,
PCI or other device IDs that they will be invoked for.

For more information about the meaning of registry fields see:
https://learn.microsoft.com/en-us/windows-hardware/drivers/hid/hidclass-hardware-ids-for-top-level-collections