mlpcre: Remove ~anchored trap
PCRE2_ANCHORED only anchors the regexp at the start (ie.  equivalent
to adding ^ at the beginning).  It *does not* anchor it at the end as
well (for which there is a separate PCRE2_ENDANCHORED which has
various traps of its own).

Everywhere I've tried to use ~anchored I've fallen into the trap of
believing it anchors both ends, causing actual bugs.  So remove it
completely from the bindings.

Replace ~anchored:true with ^...$ around the regular expression.