mldrivers: look for bootloader config in /boot/grub/grub.cfg in case of UEFI
On some older Debian-based distros (in particular, this is true for
Debian 8) GRUB config file doesn't reside on EFI partition, but rather
at /boot/grub/grub.cfg even if UEFI firmware is being used.  As a
result, in the UEFI case we get "error: no bootloader detected".  Let's
just append this path to the end of the corresponding list as a workaround
when dealing with UEFI guests.

Signed-off-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>