options: Allow --key all:SELECTOR to be used to match any device
virt-v2v (and other virt tools using --key) forces you to supply keys
on the command line per-device, eg. --key UUID:SELECTOR where UUID is
the UUID of the encrypted device.  In practice this requires users to
pre-inspect guests to find out what device UUIDs they are using (or
have some out-of-band method), match keys to UUIDs, and then pass the
result to virt-v2v.

Obviously this is a pain in the neck for users, as predicted when the
feature was initially added.  Although it is more efficient for
virt-v2v -- because attempting to decrypt a device with an invalid key
is slow -- there is a significant number of VMs (probably the vast
majority) that only have a single encrypted device, so this isn't an
issue in most cases.

This change allows --key all:SELECTOR to be used, to try the key
against any encrypted device encountered.

Fixes: https://issues.redhat.com/browse/RHEL-18142