mlcustomize: Use Start-Process -Wait to run qemu-ga installer
The qemu-ga installer is a Windows GUI program so it runs
asynchronously with the script (in Unix terms, as if the command was
run using '&').  This means that it will still be running after the
script ends, potentially causing problems with later scripts such as
the VMware uninstaller.  Plus it's generally undesirable not to wait
for it to finish.

Fixes: https://issues.redhat.com/browse/RHEL-49761
Thanks: Yan Vugenfirer
Updates: commit 5d1f5b84e44ded1c54174605810de097a339480b