mlcustomize: Add Firstboot.firstboot_dir function
This returns the name of the firstboot directory in the guest,
creating it if necessary.  We will use this to ensure we place
temporary scripts and files here rather than all over the filesystem
as before.

(cherry picked from commit 48c96e6168ec84bc6b2e9a1eea34abc257369b85)