mlcustomize: Write qemu-ga log file name to log.txt
It is apparently not possible to get the QEMU Guest Agent installer to
just log to the regular output channel, it will only write to a named
log file.  Print the name of this log file in log.txt so at least
users will know where to look for the real logs.