mlcustomize: Reboot Windows between each firstboot script
After help from some Windows friends and looking at the code in this
project: https://github.com/HCK-CI/HLK-Setup-Scripts/ we learned that
Windows installers lock files in a way that makes subsequent steps in
the firstboot process fail.

The 'fix' for this is to reboot after each stage of the install.

This leads to a large number of reboots (only the first time after
conversion), but does appear to work, even fixing VMware Tools
uninstallation which we previously thought impossible.

Fixes: https://issues.redhat.com/browse/RHEL-49761
Fixes: https://issues.redhat.com/browse/RHEL-51169
Thanks: Konstantin Kostiuk, Yuri Benditovich