mlcustomize, mltools: Check guest OS is compatible before allowing --run
If you tried to use --run/--run-command with a Windows guest
you would get the obscure error:

  virt-customize: warning: log file /builder.log: Guestfs.Error("download:
  /builder.log: No such file or directory") (ignored)
  virt-customize: error: ls: command exited with an error

Check that the guest OS is compatible with the host OS.  In fact we
can assume that the host OS is always Linux, so we're only interested
in checking that the guest OS is also Linux.

After this change, the error becomes:

  virt-customize: error: host (linux-gnu/x86_64) and guest
  (windows/i386) are not compatible, so you cannot use command line
  options that involve running commands in the guest.  Use --firstboot
  scripts instead.

This is slightly more restrictive than before, but not in a way that
matters in practice.

Also this cleans up the obscure ~incompatible_fn argument so it does
something sensible, and improves the documentation.

Fixes: https://issues.redhat.com/browse/RHEL-67560
Fixes: https://issues.redhat.com/browse/RHEL-67565
Reported-by: Ming Xie