qemuopts: Add ability to add raw, unquoted output to qemu scripts
Add qemuopts_add_raw function which can be used to add raw, unquoted
output to a qemu command line.  This can only be used when creating a
script.  Use with caution.

We use this with virt-v2v -o qemu to allow flexible command line
options to be defined in prior shell script.