mlcustomize: Drop old virtio-win dir name scraping
We track a hardcoded list of OS directory names for searching
virtio-win media. For example if we see `2k25` or `Win2025` we
know the drivers are for windows server 2025, etc.

That second naming, with a consistent `Win` prefix, is not really
relevant anymore for virtio-win.

* It hasn't been present on any ISO since at least 2015 AFAICT

* Since jan2020ish the virtio-win RPM has provided
  /usr/share/virtio-win/drivers/by-os/, which uses the shorter
  naming as found on the iso.

So we should be able to drop the `Win*` naming and continue finding
drivers on all modern ISO and RPM contents.

Note, the `Win*` naming _is_ still present under
/usr/share/virtio-win/drivers/$ARCH/. But we don't want to use that
path anyways, since it's a limited copy of what can be found under
by-os/. It's more of a historical oddity. So by dropping this,
we stop matching that dir as well, which is an added benefit.

One other note for completeness: the `Win11` naming is what the
upstream virtio-win build process spits out, but I don't think
that has any relevancy. The ISO and RPM is the what everyone consumes
AFAIK.

Signed-off-by: Cole Robinson <crobinso@redhat.com>