mlcustomize: virtio-win: clarify virtio-1.0 support
virt-v2v cacedec64072609c5203437a378aea2ab9a0e659 attempted to
determine this info from osinfo-db. But in practice, virtio_1_0
was always set to `true`:

    if osinfo-db reports zero best_driver results
      + fall back to assuming virtio_1_0 = true
    else:
      + ask osinfo-db if drivers support virtio_1_0

And from day 1 in 2020 when virtio-win added the osinfo-db
driver metadata, it has reported that virtio-1.0 is supported.

So in practice, virtio_1_0 was always `true`.

The only time we may have wanted to set that to `false` is if
creating a q35 windows VM but using really old virtio-win drivers,
from some time before 2020. I don't think that's a case we intend
to support. So this isn't something to fix anyway IMO

Related: https://github.com/libguestfs/virt-v2v/issues/75

Signed-off-by: Cole Robinson <crobinso@redhat.com>