mltools: Fix memory leak in OCaml binding of libosinfo
This failed in valgrind with the memory leak below:

==1227361== 353 (40 direct, 313 indirect) bytes in 1 blocks are definitely lost in loss record 2,664 of 3,425
==1227361==    at 0x52C0AEC: g_type_create_instance (gtype.c:1929)
==1227361==    by 0x52A58B3: g_object_new_internal.part.0 (gobject.c:2606)
==1227361==    by 0x52A6EFD: g_object_new_with_properties (gobject.c:2603)
==1227361==    by 0x52A7F30: g_object_new (gobject.c:2415)
==1227361==    by 0x121932: v2v_osinfo_os_find_os_by_short_id (libosinfo-c.c:166)
==1227361==    by 0x2DEE92: caml_c_call (in /home/rjones/d/virt-v2v/v2v/virt-v2v)
==1227361==    by 0x6F4E80F: ???
==1227361==    by 0x6F5648F: ???

Cole Robinson worked out what is happening here.

My copy of libosinfo contains no information about Windows 2025, thus
searching for short-id win2k25 returns no matches and throws a
Not_found exception.  However exceptions in OCaml C code are basically
longjmp, but written in assembly and invisible to the compiler:

  https://github.com/ocaml/ocaml/blob/d03434b711135a88883c558176a04effce4de3a1/runtime/amd64.S#L229-L235

This means they *do not* run any attribute((cleanup)) handlers,
eg. for g_autoptr.  Hence the g_autoptr locals in this function are
leaked.

We didn't realise until this point that OCaml exceptions and
attribute((cleanup)) are incompatible.  This may affect other code in
libguestfs.

The simple fix for this is to change the function so that it doesn't
throw exceptions on the happy path.  Instead we return Some os | None.

Thanks: Cole Robinson