mlcustomize/SELinux_relabel.ml: Relabel every mountpoint
It turns out that the SELinux setfiles command does not cross
mountpoint boundaries.  We therefore have to pass in a list of
mountpoints in order for setfiles to relabel everything.

For most existing guests this change is largely neutral, except that
we will now pass /sysroot/boot to setfiles (as that is almost always a
separate filesystem, and we ignored it before).

Fixes: https://issues.redhat.com/browse/RHEL-108174
Reported-by: Germano Veit Michel
Thanks: Ming Xie