mlcustomize/firstboot.ml: Use quoted string literals for firstboot
Targeted replacement of quoted string with quoted string literals
(added in OCaml 4.02) for the two firstboot scripts in this file.

In theory this should leave the binary completely identical.  I tested
this, and it doesn't, but the differences appear to be in unrelated
changes to timestamps and the symbol table.  The firstboot files are
the same.